move each example to a dedicated subdirectory
