+++ examples/BUILD.bazel
@@ -4,193 +4,6 @@ package(default_visibility = ["//visibility:public"])
 
 load("//tools:shared_library.bzl", "idto_cc_shared_library")
 
-cc_binary(
-    name = "acrobot",
-    srcs = ["acrobot.cc"],
-    data = [
-        ":acrobot.yaml",
-        "@drake//examples/acrobot:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "hopper",
-    srcs = ["hopper.cc"],
-    data = [
-        ":hopper.yaml",
-        ":models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "mini_cheetah",
-    srcs = ["mini_cheetah.cc"],
-    data = [
-        ":mini_cheetah.yaml",
-        ":models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-        "@gflags",
-    ],
-)
-
-cc_binary(
-    name = "pendulum",
-    srcs = ["pendulum.cc"],
-    data = [
-        ":pendulum.yaml",
-        "@drake//examples/pendulum:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "punyo",
-    srcs = ["punyo.cc"],
-    data = [
-        ":models",
-        ":punyo.yaml",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "dual_jaco",
-    srcs = ["dual_jaco.cc"],
-    data = [
-        ":dual_jaco.yaml",
-        ":models",
-        "@drake//manipulation/models/jaco_description:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "jaco",
-    srcs = ["jaco.cc"],
-    data = [
-        ":jaco.yaml",
-        ":models",
-        "@drake//manipulation/models/jaco_description:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "jaco_ball",
-    srcs = ["jaco_ball.cc"],
-    data = [
-        ":jaco_ball.yaml",
-        ":models",
-        "@drake//manipulation/models/jaco_description:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "kuka",
-    srcs = ["kuka.cc"],
-    data = [
-        ":kuka.yaml",
-        ":models",
-        "@drake//manipulation/models/iiwa_description:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "allegro_hand",
-    srcs = ["allegro_hand.cc"],
-    data = [
-        ":allegro_hand.yaml",
-        ":allegro_hand_upside_down.yaml",
-        ":models",
-        "@drake//manipulation/models/allegro_hand_description:models",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-        "@gflags",
-    ],
-)
-
-cc_binary(
-    name = "airhockey",
-    srcs = ["airhockey.cc"],
-    data = [
-        ":airhockey.yaml",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
-cc_binary(
-    name = "spinner",
-    srcs = ["spinner.cc"],
-    data = [
-        ":models",
-        ":spinner.yaml",
-    ],
-    deps = [
-        ":example_base",
-        "//utils:find_resource",
-        "@drake//multibody/parsing",
-        "@drake//multibody/plant",
-    ],
-)
-
 idto_cc_shared_library(
     name = "mpc_controller",
     srcs = ["mpc_controller.cc"],
@@ -256,4 +69,11 @@ idto_cc_shared_library(
     ":pd_plus_controller",
     ":yaml_config",
   ],
+)
+
+filegroup(
+    name = "models",
+    srcs = glob([
+        "models/**",
+    ]),
 )
