+++ examples/allegro_hand/allegro_hand_upside_down.yaml
