+++ examples/jaco/BUILD.bazel
@@ -1,0 +1,22 @@
+# -*- python -*-
+
+package(default_visibility = ["//visibility:public"])
+
+load("//tools:shared_library.bzl", "idto_cc_shared_library")
+
+cc_binary(
+    name = "jaco",
+    srcs = ["jaco.cc"],
+    data = [
+        ":jaco.yaml",
+        "//examples:models",
+        "@drake//manipulation/models/jaco_description:models",
+    ],
+    deps = [
+        "//examples:example_base",
+        "//utils:find_resource",
+        "@drake//multibody/parsing",
+        "@drake//multibody/plant",
+    ],
+)
+
