+++ examples/kuka/BUILD.bazel
@@ -1,0 +1,21 @@
+# -*- python -*-
+
+package(default_visibility = ["//visibility:public"])
+
+load("//tools:shared_library.bzl", "idto_cc_shared_library")
+
+cc_binary(
+    name = "kuka",
+    srcs = ["kuka.cc"],
+    data = [
+        ":kuka.yaml",
+        "//examples:models",
+        "@drake//manipulation/models/iiwa_description:models",
+    ],
+    deps = [
+        "//examples:example_base",
+        "//utils:find_resource",
+        "@drake//multibody/parsing",
+        "@drake//multibody/plant",
+    ],
+)
