+++ examples/kuka/kuka.yaml
