+++ examples/mini_cheetah/BUILD.bazel
@@ -1,0 +1,21 @@
+# -*- python -*-
+
+package(default_visibility = ["//visibility:public"])
+
+load("//tools:shared_library.bzl", "idto_cc_shared_library")
+
+cc_binary(
+    name = "mini_cheetah",
+    srcs = ["mini_cheetah.cc"],
+    data = [
+        ":mini_cheetah.yaml",
+        "//examples:models",
+    ],
+    deps = [
+        "//examples:example_base",
+        "//utils:find_resource",
+        "@drake//multibody/parsing",
+        "@drake//multibody/plant",
+        "@gflags",
+    ],
+)
