+++ examples/mini_cheetah/mini_cheetah.yaml
