+++ examples/pendulum/BUILD.bazel
@@ -1,0 +1,20 @@
+# -*- python -*-
+
+package(default_visibility = ["//visibility:public"])
+
+load("//tools:shared_library.bzl", "idto_cc_shared_library")
+
+cc_binary(
+    name = "pendulum",
+    srcs = ["pendulum.cc"],
+    data = [
+        ":pendulum.yaml",
+        "@drake//examples/pendulum:models",
+    ],
+    deps = [
+        "//examples:example_base",
+        "//utils:find_resource",
+        "@drake//multibody/parsing",
+        "@drake//multibody/plant",
+    ],
+)
