+++ examples/punyo/punyo.cc
@@ -1,5 +1,6 @@
 #include "examples/example_base.h"
 
+#include <drake/common/find_resource.h>
 #include <drake/geometry/proximity_properties.h>
 #include <drake/multibody/parsing/parser.h>
 #include <drake/multibody/plant/multibody_plant.h>
@@ -7,7 +8,7 @@
 
 namespace idto {
 namespace examples {
-namespace jaco_ball {
+namespace punyo {
 
 using drake::geometry::AddCompliantHydroelasticProperties;
 using drake::geometry::AddContactMaterial;
@@ -26,33 +27,33 @@ using drake::multibody::SpatialInertia;
 using drake::multibody::UnitInertia;
 using Eigen::Vector3d;
 
-class JacoBallExample : public TrajOptExample {
+class PunyoExample : public TrajOptExample {
  public:
-  JacoBallExample() {
+  PunyoExample() {
     // Set the camera viewpoint
-    std::vector<double> p = {1.5, 0.5, 0.0};
+    std::vector<double> p = {0.0, 1.0, -3.0};
     meshcat_->SetProperty("/Cameras/default/rotated/<object>", "position", p);
   }
 
  private:
   void CreatePlantModel(MultibodyPlant<double>* plant) const final {
-    const drake::Vector4<double> blue(0.1, 0.3, 0.5, 0.8);
-    const drake::Vector4<double> black(0.0, 0.0, 0.0, 0.5);
+    const drake::Vector4<double> blue(0.1, 0.3, 0.5, 1.0);
+    const drake::Vector4<double> green(0.3, 0.6, 0.4, 1.0);
+    const drake::Vector4<double> black(0.0, 0.0, 0.0, 1.0);
 
-    // Add a jaco arm without gravity
-    std::string robot_file = idto::FindIDTOResourceOrThrow(
-        "examples/models/j2s7s300_arm_sphere_collision_v2.sdf");
-    ModelInstanceIndex jaco = Parser(plant).AddModels(robot_file)[0];
-    RigidTransformd X_jaco(RollPitchYawd(0, 0, M_PI_2),
-                           Vector3d(0, 0.27, 0.11));
-    plant->WeldFrames(plant->world_frame(), plant->GetFrameByName("base"),
-                      X_jaco);
-    plant->set_gravity_enabled(jaco, false);
+    // Add a humanoid model
+    std::string sdf_file =
+        idto::FindIDTOResourceOrThrow("examples/models/punyoid.sdf");
+    ModelInstanceIndex humanoid = Parser(plant).AddModels(sdf_file)[0];
+    plant->WeldFrames(plant->world_frame(), plant->GetFrameByName("base"));
+    plant->set_gravity_enabled(humanoid, false);
 
-    // Add the ball model
-    const double mass = 0.3;
-    const double radius = 0.06;
+    // Add a free-floating ball to pick up
     ModelInstanceIndex ball_idx = plant->AddModelInstance("ball");
+
+    const double mass = 1.0;
+    const double radius = 0.2;
+
     const SpatialInertia<double> I(mass, Vector3d::Zero(),
                                    UnitInertia<double>::SolidSphere(radius));
     const RigidBody<double>& ball = plant->AddRigidBody("ball", ball_idx, I);
@@ -78,41 +79,30 @@ class JacoBallExample : public TrajOptExample {
                                   "ball_marker_three", black);
 
     // Add the ground
-    const drake::Vector4<double> tan(0.87, 0.7, 0.5, 1.0);
-    const drake::Vector4<double> green(0.3, 0.6, 0.4, 1.0);
-    RigidTransformd X_ground(Vector3d(0.0, 0.0, -0.5));
-    RigidTransformd X_table(Vector3d(0.6, 0.0, -0.499));
-    plant->RegisterVisualGeometry(plant->world_body(), X_ground, Box(25, 25, 1),
-                                  "ground", green);
-    plant->RegisterVisualGeometry(plant->world_body(), X_table,
-                                  Box(1.5, 1.5, 1), "table", tan);
+    RigidTransformd X_ground(Vector3d(0.0, 0.0, -5.0));
+    plant->RegisterVisualGeometry(plant->world_body(), X_ground,
+                                  Box(25, 25, 10), "ground", green);
     plant->RegisterCollisionGeometry(plant->world_body(), X_ground,
-                                     Box(25, 25, 1), "ground",
+                                     Box(25, 25, 10), "ground",
                                      CoulombFriction<double>(0.5, 0.5));
   }
 
   void CreatePlantModelForSimulation(
       MultibodyPlant<double>* plant) const final {
-    const drake::Vector4<double> blue(0.1, 0.3, 0.5, 0.8);
-    const drake::Vector4<double> black(0.0, 0.0, 0.0, 0.5);
+    const drake::Vector4<double> blue(0.1, 0.3, 0.5, 1.0);
+    const drake::Vector4<double> green(0.3, 0.6, 0.4, 1.0);
+    const drake::Vector4<double> black(0.0, 0.0, 0.0, 1.0);
 
-    // Add a jaco arm without gravity
-    std::string robot_file = idto::FindIDTOResourceOrThrow(
-        "examples/models/j2s7s300_arm_hydro_collision.sdf");
-    ModelInstanceIndex jaco = Parser(plant).AddModels(robot_file)[0];
-    RigidTransformd X_jaco(RollPitchYawd(0, 0, M_PI_2),
-                           Vector3d(0, 0.27, 0.11));
-    plant->WeldFrames(plant->world_frame(), plant->GetFrameByName("base"),
-                      X_jaco);
-    plant->set_gravity_enabled(jaco, false);
+    // Add a humanoid model
+    std::string sdf_file =
+        FindIDTOResourceOrThrow("examples/models/punyoid.sdf");
+    ModelInstanceIndex humanoid = Parser(plant).AddModels(sdf_file)[0];
+    plant->WeldFrames(plant->world_frame(), plant->GetFrameByName("base"));
+    plant->set_gravity_enabled(humanoid, false);
 
-    // Add the ball model
-    // N.B. the radius is slightly larger than the model used for optimization,
-    // in order to (1) compensate for the fact that the jaco arm cannot apply
-    // feed-forward torques and (2) compensate for the fact that the planner's
-    // contact model allows force at a distance.
-    const double mass = 0.3;
-    const double radius = 0.063;
+    // Add a free-floating ball to pick up
+    const double mass = 1.0;
+    const double radius = 0.2;
     ModelInstanceIndex ball_idx = plant->AddModelInstance("ball");
     const SpatialInertia<double> I(mass, Vector3d::Zero(),
                                    UnitInertia<double>::SolidSphere(radius));
@@ -120,9 +110,14 @@ class JacoBallExample : public TrajOptExample {
 
     plant->RegisterVisualGeometry(ball, RigidTransformd::Identity(),
                                   Sphere(radius), "ball_visual", blue);
+
+    ProximityProperties ball_proximity;
+    AddContactMaterial(3.0, {}, CoulombFriction<double>(1.5, 1.5),
+                       &ball_proximity);
+    AddCompliantHydroelasticProperties(0.1, 1e5, &ball_proximity);
     plant->RegisterCollisionGeometry(ball, RigidTransformd::Identity(),
                                      Sphere(radius), "ball_collision",
-                                     CoulombFriction<double>(0.5, 0.5));
+                                     ball_proximity);
 
     // Add some markers to the ball so we can see its rotation
     RigidTransformd X_m1(RollPitchYawd(0, 0, 0), Vector3d(0, 0, 0));
@@ -139,26 +134,21 @@ class JacoBallExample : public TrajOptExample {
                                   "ball_marker_three", black);
 
     // Add the ground
-    const drake::Vector4<double> tan(0.87, 0.7, 0.5, 1.0);
-    const drake::Vector4<double> green(0.3, 0.6, 0.4, 1.0);
-    RigidTransformd X_ground(Vector3d(0.0, 0.0, -0.5));
-    RigidTransformd X_table(Vector3d(0.6, 0.0, -0.499));
-    plant->RegisterVisualGeometry(plant->world_body(), X_ground, Box(25, 25, 1),
-                                  "ground", green);
-    plant->RegisterVisualGeometry(plant->world_body(), X_table,
-                                  Box(1.5, 1.5, 1), "table", tan);
+    RigidTransformd X_ground(Vector3d(0.0, 0.0, -5.0));
+    plant->RegisterVisualGeometry(plant->world_body(), X_ground,
+                                  Box(25, 25, 10), "ground", green);
     plant->RegisterCollisionGeometry(plant->world_body(), X_ground,
-                                     Box(25, 25, 1), "ground",
+                                     Box(25, 25, 10), "ground",
                                      CoulombFriction<double>(0.5, 0.5));
   }
 };
 
-}  // namespace jaco_ball
+}  // namespace punyo
 }  // namespace examples
 }  // namespace idto
 
 int main() {
-  idto::examples::jaco_ball::JacoBallExample example;
-  example.RunExample("examples/jaco_ball.yaml");
+  idto::examples::punyo::PunyoExample example;
+  example.RunExample("examples/punyo/punyo.yaml");
   return 0;
 }
