+++ examples/spinner/spinner.yaml
